# im-services/docs/swagger.yaml
basePath: /api
definitions:
  auth.loginResponse:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱账号
        type: string
      expire_time:
        description: token过期时间
        type: integer
      id:
        description: 用户id
        type: integer
      name:
        description: 名称
        type: string
      token:
        description: token
        type: string
      ttl:
        description: token有效期
        type: integer
      uid:
        description: uid
        type: string
    type: object
  cloud.Response:
    properties:
      file_url:
        type: string
    type: object
...
  response.JsonResponse:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 返回结构数据
      httpCode:
        description: http状态码
        type: integer
      message:
        description: 消息内容
        type: string
    type: object
...
host: 127.0.0.1:8000
info:
  contact:
    email: pltrueover@gmail.com
    name: im-services
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: im-services  接口文档
  version: "1.0"
paths:
  /auth/githubLogin:
    get:
      description: github登录
      parameters:
      - description: github授权码
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.loginResponse'
              type: object
      summary: friends github登录
      tags:
      - 登录相关
  /auth/login:
    post:
      consumes:
      - multipart/form-data
      description: 登录接口
      parameters:
      - description: 账号
        in: formData
        name: email
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.loginResponse'
              type: object
      summary: Login 登录
      tags:
      - 登录相关
  /auth/registered:
    post:
      consumes:
      - multipart/form-data
      description: 注册接口
      parameters:
      - description: 账号
        in: formData
        name: email
        required: true
        type: string
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      - description: 邮件类型 1.注册 2.找回密码
        in: formData
        name: email_type
        type: integer
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 确认密码
        in: formData
        name: password_repeat
        required: true
        type: string
      - description: 验证码
        in: formData
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: Registered 注册
      tags:
      - 登录相关
  /auth/sendEmailCode:
    post:
      consumes:
      - multipart/form-data
      description: 发送邮件接口
      parameters:
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 邮件类型 1.注册 2.找回密码
        in: formData
        name: email_type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: Registered 发送邮件
      tags:
      - 登录相关
...
  /upload/file:
    post:
      description: 文件上传接口
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/cloud.Response'
              type: object
      summary: upload/file 文件上传接口
      tags:
      - 文件
...
swagger: "2.0"